+++ packages/gatsby-plugin-image/gatsby-browser.js
@@ -1,0 +1,11 @@
+const React = require("react")
+
+const { LaterHydrator } = require(".")
+
+exports.wrapRootElement = ({ element }) => {
+  return (
+    <LaterHydrator>
+      {element}
+    </LaterHydrator>
+  )
+}